move all logic from loop to agent file
