+++ computer-use-demo/computer_use_demo/agents/agent.py
@@ -2,25 +2,25 @@
 Agent implementation for Computer Use Demo.
 """
 
-from typing import Any, Callable, List, Optional, cast
+import platform
+from datetime import datetime
+from enum import StrEnum
+from typing import Any, Callable, Dict, List, Optional, cast
 
 import httpx
 from anthropic import APIError, APIResponseValidationError, APIStatusError
 from anthropic.types.beta import (
+    BetaCacheControlEphemeralParam,
     BetaContentBlockParam,
+    BetaImageBlockParam,
+    BetaMessage,
     BetaMessageParam,
+    BetaTextBlock,
     BetaTextBlockParam,
     BetaToolResultBlockParam,
+    BetaToolUseBlockParam,
 )
 
-from computer_use_demo.loop import (
-    PROMPT_CACHING_BETA_FLAG,
-    SYSTEM_PROMPT,
-    APIProvider,
-    _inject_prompt_caching,
-    _make_api_tool_result,
-    _response_to_params,
-)
 from computer_use_demo.tools import (
     TOOL_GROUPS_BY_VERSION,
     ToolCollection,
@@ -33,6 +33,128 @@ from .logging import get_logger
 
 # Get logger for this module
 logger = get_logger("computer_use_demo.agents.agent")
+
+PROMPT_CACHING_BETA_FLAG = "prompt-caching-2024-07-31"
+
+
+class APIProvider(StrEnum):
+    ANTHROPIC = "anthropic"
+    BEDROCK = "bedrock"
+    VERTEX = "vertex"
+
+
+# This system prompt is optimized for the Docker environment in this repository and
+# specific tool combinations enabled.
+# We encourage modifying this system prompt to ensure the model has context for the
+# environment it is running in, and to provide any additional information that may be
+# helpful for the task at hand.
+SYSTEM_PROMPT = f"""<SYSTEM_CAPABILITY>
+* You are utilising an Ubuntu virtual machine using {platform.machine()} architecture with internet access.
+* You can feel free to install Ubuntu applications with your bash tool. Use curl instead of wget.
+* To open firefox, please just click on the firefox icon.  Note, firefox-esr is what is installed on your system.
+* Using bash tool you can start GUI applications, but you need to set export DISPLAY=:1 and use a subshell. For example "(DISPLAY=:1 xterm &)". GUI apps run with bash tool will appear within your desktop environment, but they may take some time to appear. Take a screenshot to confirm it did.
+* When using your bash tool with commands that are expected to output very large quantities of text, redirect into a tmp file and use str_replace_editor or `grep -n -B <lines before> -A <lines after> <query> <filename>` to confirm output.
+* When viewing a page it can be helpful to zoom out so that you can see everything on the page.  Either that, or make sure you scroll down to see everything before deciding something isn't available.
+* When using your computer function calls, they take a while to run and send back to you.  Where possible/feasible, try to chain multiple of these calls all into one function calls request.
+* The current date is {datetime.today().strftime('%A, %B %-d, %Y')}.
+</SYSTEM_CAPABILITY>
+
+<IMPORTANT>
+* When using Firefox, if a startup wizard appears, IGNORE IT.  Do not even click "skip this step".  Instead, click on the address bar where it says "Search or enter address", and enter the appropriate search term or URL there.
+* If the item you are looking at is a pdf, if after taking a single screenshot of the pdf it seems that you want to read the entire document instead of trying to continue to read the pdf from your screenshots + navigation, determine the URL, use curl to download the pdf, install and use pdftotext to convert it to a text file, and then read that text file directly with your StrReplaceEditTool.
+</IMPORTANT>"""
+
+
+def response_to_params(
+    response: BetaMessage,
+) -> list[BetaContentBlockParam]:
+    """Convert a BetaMessage to a list of BetaContentBlockParam."""
+    res: list[BetaContentBlockParam] = []
+    for block in response.content:
+        if isinstance(block, BetaTextBlock):
+            if block.text:
+                res.append(BetaTextBlockParam(type="text", text=block.text))
+            elif getattr(block, "type", None) == "thinking":
+                # Handle thinking blocks - include signature field
+                thinking_block = {
+                    "type": "thinking",
+                    "thinking": getattr(block, "thinking", None),
+                }
+                if hasattr(block, "signature"):
+                    thinking_block["signature"] = getattr(block, "signature", None)
+                res.append(cast(BetaContentBlockParam, thinking_block))
+        else:
+            # Handle tool use blocks normally
+            res.append(cast(BetaToolUseBlockParam, block.model_dump()))
+    return res
+
+
+def inject_prompt_caching(
+    messages: list[BetaMessageParam],
+):
+    """
+    Set cache breakpoints for the 3 most recent turns
+    one cache breakpoint is left for tools/system prompt, to be shared across sessions
+    """
+    breakpoints_remaining = 3
+    for message in reversed(messages):
+        if message["role"] == "user" and isinstance(
+            content := message["content"], list
+        ):
+            if breakpoints_remaining:
+                breakpoints_remaining -= 1
+                # Use type ignore to bypass TypedDict check until SDK types are updated
+                content[-1]["cache_control"] = BetaCacheControlEphemeralParam(  # type: ignore
+                    {"type": "ephemeral"}
+                )
+            else:
+                # Add type cast to avoid Pyright error on pop operation
+                cast(Dict[str, Any], content[-1]).pop("cache_control", None)
+                # we'll only every have one extra turn per loop
+                break
+
+
+def maybe_prepend_system_tool_result(result: ToolResult, result_text: str) -> str:
+    """Prepend system information to tool result if available."""
+    if result.system:
+        result_text = f"<s>{result.system}</s>\n{result_text}"
+    return result_text
+
+
+def make_api_tool_result(
+    result: ToolResult, tool_use_id: str
+) -> BetaToolResultBlockParam:
+    """Convert an agent ToolResult to an API ToolResultBlockParam."""
+    tool_result_content: list[BetaTextBlockParam | BetaImageBlockParam] | str = []
+    is_error = False
+    if result.error:
+        is_error = True
+        tool_result_content = maybe_prepend_system_tool_result(result, result.error)
+    else:
+        if result.output:
+            tool_result_content.append(
+                {
+                    "type": "text",
+                    "text": maybe_prepend_system_tool_result(result, result.output),
+                }
+            )
+        if result.base64_image:
+            tool_result_content.append(
+                {
+                    "type": "image",
+                    "source": {
+                        "type": "base64",
+                        "media_type": "image/png",
+                        "data": result.base64_image,
+                    },
+                }
+            )
+    return {
+        "type": "tool_result",
+        "content": tool_result_content,
+        "tool_use_id": tool_use_id,
+        "is_error": is_error,
+    }
 
 
 class Agent:
@@ -157,7 +279,7 @@ class Agent:
             # Setup prompt caching if enabled
             if enable_prompt_caching:
                 betas.append(PROMPT_CACHING_BETA_FLAG)
-                _inject_prompt_caching(self.history.messages)
+                inject_prompt_caching(self.history.messages)
                 # Because cached reads are 10% of the price, we don't think it's
                 # ever sensible to break the cache by truncating images
                 only_n_most_recent_images = 0
@@ -216,7 +338,7 @@ class Agent:
             response = raw_response.parse()
 
             # Process the response
-            response_params = _response_to_params(response)
+            response_params = response_to_params(response)
             self.history.append(
                 {
                     "role": "assistant",
@@ -238,7 +360,7 @@ class Agent:
                         tool_input=cast(dict[str, Any], content_block["input"]),
                     )
                     tool_result_content.append(
-                        _make_api_tool_result(result, content_block["id"])
+                        make_api_tool_result(result, content_block["id"])
                     )
                     tool_output_callback(result, content_block["id"])
                     logger.debug(
